refactor: optimize framework structure and browser support
- Reorganized page objects into sub-packages (Admin/Recruitment)
- Integrated WebDriverManager to fix driver binary errors
- Added support for Brave browser with binary path configuration
- Updated pom.xml with latest stable dependencies
